+++ src/background.ts
@@ -1,0 +1,5 @@
+browser.runtime.onInstalled.addListener(function (details) {
+    if (details.reason == "install") {
+        browser.storage.sync.set({ displayOptionsButton: true });
+    }
+});